backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
beautifulsoup4==4.14.2
black==24.10.0
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.2.1
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distlib==0.4.0
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.13.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.8.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-extra-fields==3.7.0
drf-spectacular==0.27.2
filelock==3.19.1
filetype==1.2.0
flake8==7.3.0
google==3.0.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
hyperlink==21.0.0
identify==2.6.13
idna==3.11
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
isort==5.13.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.8.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==7.0.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
whitenoise==6.11.0
zope.interface==8.0.1
